requirement.txt
django-ckeditor
django-environ